Maximum Edge
Given N number of vertices, M number of edges in an undirected graph,task is to find the maximum number of edges among all the connected components in the given graph. The total number of edges in a connected component of an undirected graph is equal to half of the total sum of the degrees of all of its vertices.